moved department pipe to shared
