+++ src/app/app.module.ts
@@ -14,7 +14,7 @@ import {ListsService} from './lists/lists.service';
 import {ItemsComponent} from './items/items.component';
 import {ItemService} from './items/item.service';
 import {DepartmentService} from "./shared/department.service";
-import { DepartmentPipe } from './items/department.pipe';
+import { DepartmentPipe } from './shared/department.pipe';
 
 const appRoutes: Routes = [
   {path: 'lists', component: ListsComponent},
